Handling Errors

+++ backend/routes/post.js
@@ -40,17 +40,24 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.user.userId
     });
-    post.save().then(createdPost => {
-      res.status(201).json({
-        message: "Post added successfully",
-        post: {
-          ...createdPost,
-          id: createdPost._id
-        }
-      });
-    });
+    post.save()
+      .then(createdPost => {
+        res.status(201).json({
+          message: "Post added successfully",
+          post: {
+            ...createdPost,
+            id: createdPost._id
+          }
+        })
+      })
+      .catch(err => {
+        res.status(500).json({
+          message: "Creating a new post failed!"
+        })
+      })
   }
 );
 
@@ -68,11 +75,21 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.user.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
-    });
+    Post.updateOne({ _id: req.params.id, creator: req.user.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      } else {
+        res.status(401).json({ message: "Not Authorized!" });
+      }
+    })
+      .catch(err => {
+        res.status(500).json({
+          message: "Updating a new post failed!"
+        })
+      })
   }
 );
 
@@ -109,9 +126,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.user.userId }).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   });
 });
